requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.3
django-bootstrap4==2.3.1
django-chartjs==2.2.1
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1
